requirements.txt
crewai
crewai_tools
langchain_community
langchain_openai
flask
python-dotenv
